refactor(requests): change the directory into subfolders
