+++ backend/app/Http/Requests/Features/TargetRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Features;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class LogRequest extends FormRequest
+class TargetRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,8 +22,10 @@ class LogRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'action' => 'required|string|max:255',
-            'details' => 'nullable|string',
+            'name' => 'required|string|max:255',
+            'target_amount' => 'required|numeric|min:0',
+            'saved_amount' => 'required|numeric|min:0',
+            'deadline' => 'required|date',
         ];
     }
 }
